Сделал сервисы для всех моделей, изменил все контроллеры под сервисы.Добавил валидацию

+++ app/services/columns/delete_column_service.rb
@@ -1,18 +1,16 @@
 class Columns::DeleteColumnService < ApplicationService
-    attr_reader :id
+  attr_reader :id
 
-    def initialize(id)
-        @id = id
-    end
+  def initialize(id)
+    @id = id
+  end
 
-    def call
-        column = Column.find(@id)
-        column.destroy
-        success = true
-      rescue => errors
-        succes = false
-      ensure
-        return OpenStruct.new(success?: success, column: column, errors: errors)
-      end
+  def call
+    column = Column.find_by(id: id)
 
+    return OpenStruct.new(success?: false, column: nil, errors: ['Column not found']) if column.blank?
+
+    column.destroy
+    OpenStruct.new(success?: column.destroyed?, column: column, errors: column.errors.full_messages)
+  end
 end
